이미지 null값 처리

+++ backend/src/controllers/mypageController.js
@@ -130,13 +130,19 @@ class UserController {
   }
 
   async DeleteUserImage(req, res) {
+
     const before_image_path = req.body.preimage;
-    const image_path = path.join(__dirname,'../../../../front/front/front/public/', `${before_image_path}`);
+
+    if(!before_image_path){
+      res.status(200).send({access: true});
+      return;
+    }else{
+      const image_path = path.join(__dirname,'../../../../front/front/front/public/', `${before_image_path}`);
     
-    await fs.unlink(image_path);
-
-    res.status(200).send({access: true})
-
+      await fs.unlink(image_path);
+  
+      res.status(200).send({access: true})
+    }
   }
 }
